+++ packages/react-native-editor/__device-tests__/gutenberg-editor-media-blocks-@canary.test.js
@@ -125,9 +125,9 @@ onlyOniOS( 'Gutenberg Editor Cover Block test', () => {
 
 		const { height } = await coverBlock.getSize();
 		// Height is set to 20rem, where 1rem is 16.
-		// There is also block's vertical padding equal 32.
-		// Finally, the total height should be 20 * 16 + 32 = 352.
-		expect( height ).toBe( 352 );
+		// There is also block's vertical padding equal 16.
+		// Finally, the total height should be 20 * 16 + 16 = 336.
+		expect( height ).toBe( 336 );
 
 		await coverBlock.click();
 		expect( coverBlock ).toBeTruthy();