+++ packages/react-native-editor/CHANGELOG.md
@@ -10,11 +10,12 @@ For each user feature we should also add a importance categorization label  to i
 -->
 
 ## Unreleased
+-   [**] Add media inserter buttons to editor toolbar [#51827]
+-   [**] Update native BlockOutline component styles to remove blue border from blocks [#51222]
 
 ## 1.99.0
 -   [*] Rename "Reusable blocks" to "Synced patterns", aligning with the web editor. [#51704]
 -   [**] Fix a crash related to Reanimated when closing the editor [#52320]
--   [**] Add media inserter buttons to editor toolbar [#51827]
 
 ## 1.98.1
 -   [*] fix: Display heading level dropdown icons and labels [#52004]